// src/app/models/report.model.ts
export interface LeadReport {
  totalProspects: number;
  totalInteractions: number;
  totalInteractionsByPhone: number;
  totalInteractionsByEmail: number;
  totalInteractionsByPhoneTrue: number;
  totalInteractionsByPhoneFalse: number;
  totalInteractionsByEmailTrue: number;
  totalInteractionsByEmailFalse: number;
  totalProspectContacted: number;
  totalProspectContactedTrue: number;
}